# Faster sse4 and avx2 SIMD blitters