I've improved avx2 and sse4 SIMD blit blending code. It is now **~1.5x faster**.
New code calculates exactly same byte values as before.

To measure performance I called SDL_BlitSurface for 256x256 ABGR surface 1000x in a loop.
Here are timings for old -> new code on AMD and Intel CPU's on Windows using MSVC v19.41.

AVX2 blit:
```
Ryzen 5950x: 21.6 msec -> 14.2 msec = 1.52x faster
i7-7940X:    24.6 msec -> 17.5 msec = 1.41x faster
```

SSE4 blit:
```
Ryzen 5950x: 44.1 msec -> 27.8 msec = 1.59x faster
i7-7940X:    46.5 msec -> 29.9 msec = 1.55x faster
```

I have run `testautomation.exe --filter Blending` on both codepaths to check that hashes of blit results matches expected values.